third_party/blink/web_tests/platform/generic/accessibility/not-ignore-landmark-roles-expected.txt

article element
section element
header element
aside element
address element
footer element
main element
nav element

This tests that landmark and related roles are not ignored.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS containerNode.childrenCount is 8
PASS AXObject.role is 'AXRole: AXArticle'
PASS AXObject.role is 'AXRole: AXSection'
PASS AXObject.role is 'AXRole: AXHeader'
PASS AXObject.role is 'AXRole: AXComplementary'
PASS AXObject.role is 'AXRole: AXGenericContainer'
PASS AXObject.role is 'AXRole: AXFooter'
PASS AXObject.role is 'AXRole: AXMain'
PASS AXObject.role is 'AXRole: AXNavigation'
PASS successfullyParsed is true

TEST COMPLETE
